requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.12
    # via litellm
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via markdownify
blis==1.3.0
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   duckduckgo-search
    #   litellm
    #   typer
cloudpathlib==0.21.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
distro==1.9.0
    # via openai
duckduckgo-search==8.0.4
    # via smolagents
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.1.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   litellm
    #   ollama
    #   openai
huggingface-hub==0.33.0
    # via
    #   smolagents
    #   tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via
    #   litellm
    #   smolagents
    #   spacy
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.0
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
litellm==1.72.4
    # via syntheticgenagent (pyproject.toml)
lxml==5.4.0
    # via duckduckgo-search
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markdownify==1.1.0
    # via smolagents
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.3.0
    # via
    #   syntheticgenagent (pyproject.toml)
    #   blis
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
ollama==0.5.1
    # via syntheticgenagent (pyproject.toml)
openai==1.86.0
    # via litellm
packaging==25.0
    # via
    #   huggingface-hub
    #   spacy
    #   thinc
    #   weasel
pillow==11.2.1
    # via smolagents
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.8.0/pl_core_news_sm-3.8.0-py3-none-any.whl
    # via syntheticgenagent (pyproject.toml)
preshed==3.0.10
    # via
    #   spacy
    #   thinc
primp==0.15.0
    # via duckduckgo-search
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pyaml==25.5.0
    # via syntheticgenagent (pyproject.toml)
pydantic==2.11.5
    # via
    #   confection
    #   litellm
    #   ollama
    #   openai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via
    #   syntheticgenagent (pyproject.toml)
    #   litellm
    #   smolagents
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   pyaml
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.4
    # via
    #   huggingface-hub
    #   smolagents
    #   spacy
    #   tiktoken
    #   weasel
rich==14.0.0
    # via
    #   smolagents
    #   typer
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.7.0
    # via syntheticgenagent (pyproject.toml)
scipy==1.15.3
    # via scikit-learn
setuptools==80.9.0
    # via
    #   marisa-trie
    #   spacy
    #   thinc
shellingham==1.5.4
    # via typer
six==1.17.0
    # via markdownify
smart-open==7.1.0
    # via weasel
smolagents==1.18.0
    # via syntheticgenagent (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
spacy==3.8.7
    # via syntheticgenagent (pyproject.toml)
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.6
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.1
    # via litellm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   spacy
typer==0.16.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.14.0
    # via
    #   anyio
    #   beautifulsoup4
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.17.2
    # via smart-open
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
